:tada: feat: ProtocoloFollow concluído e inserção de paginação nos protocolos

+++ Controllers/ProtocoloController.cs
@@ -3,25 +3,47 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using TesteDevDbm.Context;
 using TesteDevDbm.Models;
+using TesteDevDbm.Models.ViewModels;
+using TesteDevDbm.Services;
 
 namespace TesteDevDbm.Controllers
 {
     public class ProtocoloController : Controller
     {
         private readonly ProtocoloContext _context;
+        private readonly IProtocoloFollowService _protocoloFollowService;
 
-        public ProtocoloController(ProtocoloContext context)
+        public ProtocoloController(ProtocoloContext context, IProtocoloFollowService protocoloFollowService)
         {
             _context = context;
+            _protocoloFollowService = protocoloFollowService;
         }
-        public IActionResult Index()
+        public IActionResult Index(int pageNumber = 1, int pageSize = 5)
         {
             var protocolos = _context.Protocolos
                 .Include(p => p.Cliente)
-                .Include(p => p.ProtocoloStatus).ToList();
-            return View(protocolos);
-        }
+                .Include(p => p.ProtocoloStatus);
 
+            // Calcular o total de registros
+            var totalRecords = protocolos.Count();
+
+            // Paginando os registros
+            var protocolosPaginated = protocolos
+                .Skip((pageNumber - 1) * pageSize)
+                .Take(pageSize)
+                .ToList();
+
+            // Criar um objeto para passar para a view
+            var viewModel = new ProtocoloViewModel
+            {
+                Protocolos = protocolosPaginated,
+                CurrentPage = pageNumber,
+                TotalPages = (int)Math.Ceiling((double)totalRecords / pageSize)
+            };
+
+            return View(viewModel);
+        }
+        
         [HttpGet]
         [Route("Protocolo/Criar")]
         public IActionResult Criar()
@@ -39,6 +61,13 @@ namespace TesteDevDbm.Controllers
             {
                 _context.Protocolos.Add(protocolo);
                 _context.SaveChanges();
+
+                var protocoloFollow = new ProtocoloFollow();
+                protocoloFollow.ProtocoloId = protocolo.IdProtocolo;
+                protocoloFollow.DataAcao = protocolo.DataAbertura;
+                protocoloFollow.DescricaoAcao = "Criação do protocolo";
+
+                _protocoloFollowService.CriaProtocoloFollow(protocoloFollow);
                 return RedirectToAction(nameof(Index));
             }
             return View(protocolo);
@@ -58,7 +87,7 @@ namespace TesteDevDbm.Controllers
                 return NotFound();
             }
 
-            ViewBag.Clientes =new SelectList(_context.Clientes.ToList(), "IdCliente", "Nome");
+            ViewBag.Clientes = new SelectList(_context.Clientes.ToList(), "IdCliente", "Nome");
             ViewBag.Status = new SelectList(_context.StatusProtocolos.ToList(), "IdStatus", "NomeStatus");
 
             return View(protocolo);
@@ -77,10 +106,17 @@ namespace TesteDevDbm.Controllers
 
                 _context.Update(protocolo);
                 _context.SaveChanges();
+
+                var protocoloFollow = new ProtocoloFollow();
+                protocoloFollow.ProtocoloId = protocolo.IdProtocolo;
+                protocoloFollow.DataAcao = DateOnly.FromDateTime(DateTime.Now);
+                protocoloFollow.DescricaoAcao = "Edição do protocolo";
+
+                _protocoloFollowService.CriaProtocoloFollow(protocoloFollow);
                 return RedirectToAction(nameof(Index));
             }
 
-            ViewBag.Clientes =new SelectList(_context.Clientes.ToList(), "IdCliente", "Nome");
+            ViewBag.Clientes = new SelectList(_context.Clientes.ToList(), "IdCliente", "Nome");
             ViewBag.Status = new SelectList(_context.StatusProtocolos.ToList(), "IdStatus", "NomeStatus");
             return View(protocolo);
         }
@@ -93,7 +129,7 @@ namespace TesteDevDbm.Controllers
             .FirstOrDefault(p => p.IdProtocolo == id);
 
             if (protocolo == null)
-              return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(Index));
 
             return View(protocolo);
         }
@@ -107,7 +143,7 @@ namespace TesteDevDbm.Controllers
 
 
             if (protocolo == null)
-              return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(Index));
 
             return View(protocolo);
         }
@@ -117,8 +153,12 @@ namespace TesteDevDbm.Controllers
         {
             var protocoloBanco = _context.Protocolos.Find(protocolo.IdProtocolo);
 
-            _context.Protocolos.Remove(protocoloBanco);
-            _context.SaveChanges();
+            if (protocoloBanco != null)
+            {
+                var IdProtocolo = protocoloBanco.IdProtocolo;
+                _context.Protocolos.Remove(protocoloBanco);
+                _context.SaveChanges();
+            }
 
             return RedirectToAction(nameof(Index));
         }
